+++ CHANGES.md
@@ -3,7 +3,10 @@
 0.1.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add spot price check cli.
+- Add spot BidPrice.
+- Show estimated cost.
+- Filter by tag for list cli.
 
 
 0.1.1 (2019-02-21)